api/requirements.txt
tensorflow
pillow
numpy
fastapi
uvicorn
python-multipart
python-dotenv
cors